remove qgemm_kernel_type* and rename protocol
