+++ onnxruntime/core/mlas/lib/qgemm_kernel_avx2.cpp
@@ -16,8 +16,7 @@ Abstract:
 
 #include "mlasi.h"
 #include "qgemm_dispatcher.h"
-#include "qgemm_kernel_type.h"
-#include "qgemm_kernel_protocol.h"
+#include "qgemm_kernel.h"
 
 #if defined(MLAS_TARGET_AMD64)
 
@@ -79,6 +78,21 @@ extern "C" {
         int32_t* ColumnSumBuffer
         );
 }
+
+struct MLAS_GEMM_U8S8_KERNEL_AVX2
+{
+    typedef uint8_t PackedAType;
+    typedef uint8_t PackedBType;
+    typedef int8_t OffsetBType;
+
+    static constexpr size_t PackedK = 4;
+    static constexpr MLAS_GEMM_U8X8_STRIDES Strides{ 24, 256, 128 };
+    static constexpr MLAS_GEMM_U8X8_STRIDES PackedStrides{ 48, 256, 384 };
+};
+
+constexpr size_t MLAS_GEMM_U8S8_KERNEL_AVX2::PackedK;
+constexpr MLAS_GEMM_U8X8_STRIDES MLAS_GEMM_U8S8_KERNEL_AVX2::Strides;
+constexpr MLAS_GEMM_U8X8_STRIDES MLAS_GEMM_U8S8_KERNEL_AVX2::PackedStrides;
 
 template<>
 MLAS_FORCEINLINE
@@ -175,6 +189,22 @@ const MLAS_GEMM_U8X8_DISPATCH MlasGemmU8S8DispatchAvx2 = {
     MLAS_GEMM_U8S8_KERNEL_AVX2::PackedK,
     MLAS_GEMM_U8S8_KERNEL_AVX2::PackedStrides.K,
 };
+
+struct MLAS_GEMM_U8U8_KERNEL_AVX2
+{
+    typedef int16_t PackedAType;
+    typedef uint8_t PackedBType;
+    typedef uint8_t OffsetBType;
+
+    static constexpr size_t PackedK = 2;
+    static constexpr MLAS_GEMM_U8X8_STRIDES Strides{ 24, 256, 128 };
+    static constexpr MLAS_GEMM_U8X8_STRIDES PackedStrides{ 48, 256, 384 };
+};
+
+constexpr size_t MLAS_GEMM_U8U8_KERNEL_AVX2::PackedK;
+constexpr MLAS_GEMM_U8X8_STRIDES MLAS_GEMM_U8U8_KERNEL_AVX2::Strides;
+constexpr MLAS_GEMM_U8X8_STRIDES MLAS_GEMM_U8U8_KERNEL_AVX2::PackedStrides;
+
 
 template<>
 MLAS_FORCEINLINE
