+++ onnxruntime/core/mlas/lib/qgemm_kernel_default.cpp
@@ -16,8 +16,22 @@ Abstract:
 
 #include "mlasi.h"
 #include "qgemm_dispatcher.h"
-#include "qgemm_kernel_type.h"
-#include "qgemm_kernel_protocol.h"
+#include "qgemm_kernel.h"
+
+struct MLAS_GEMM_U8X8_KERNEL_DEFAULT
+{
+    typedef uint8_t PackedAType;
+    typedef uint8_t PackedBType;
+    typedef uint8_t OffsetBType;
+
+    static constexpr size_t PackedK = 4;
+    static constexpr MLAS_GEMM_U8X8_STRIDES Strides{ 16, 128, 128 };
+    static constexpr MLAS_GEMM_U8X8_STRIDES PackedStrides{ 16, 128, 128 };
+};
+
+constexpr size_t MLAS_GEMM_U8X8_KERNEL_DEFAULT::PackedK;
+constexpr MLAS_GEMM_U8X8_STRIDES MLAS_GEMM_U8X8_KERNEL_DEFAULT::Strides;
+constexpr MLAS_GEMM_U8X8_STRIDES MLAS_GEMM_U8X8_KERNEL_DEFAULT::PackedStrides;
 
 template<>
 MLAS_FORCEINLINE
